Deployment/Backend/requirements.txt
# Backend-specific dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
apscheduler>=3.10.0
watchdog>=3.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Phase 2 dependencies
yfinance>=0.2.0
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
feedparser>=6.0.0

# Phase 3 dependencies
groq>=0.4.0

# Install fx_alphalab as editable package
-e ../../fx_alphalab
